feat: implemented WaitGroup and Defer golang sync patterns
